requirements.txt
torch>=2.0.0
safetensors>=0.3.0
PyYAML>=6.0
numpy>=1.24.0
tqdm>=4.65.0
psutil>=5.9.0
packaging>=21.0
